Update README.md
I am a big fan of Monty Python and I am always pleased to find references to the group in the Python books and lessons I read. I was reminded of this scene in Monty Python and the Holy Grail while I was learning about if statements and loops. I decided to try my hand at recreating this scene to practice. 

I think as I learn new skills I will be able to improve this project to make it more user friendly. That's something I am enjoying about coding; there is always opportunity to improve!